Correção - Ajustar respostas dos endpoints

+++ src/main/java/com/example/Medico/services/ConsultaService.java
@@ -1,6 +1,7 @@
 package com.example.Medico.services;
 
 import com.example.Medico.dtos.ConsultaDTO;
+import com.example.Medico.exceptions.ConsultaNotFoundException;
 import com.example.Medico.mappers.ConsultaMapper;
 import com.example.Medico.models.Consulta;
 import com.example.Medico.repository.ConsultaRepository;
@@ -11,8 +12,6 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
-import java.util.Optional;
-
 @Service
 public class ConsultaService {
 
@@ -20,15 +19,18 @@ public class ConsultaService {
     private ConsultaRepository consultaRepository;
 
     // Buscar todas as consultas disponíveis
+    
     public Page<ConsultaDTO> findAll(Pageable pageable) {
         return consultaRepository.findAll(pageable)
                 .map(ConsultaMapper::convertToDTO);
     }
 
     // Buscar individualmente a consulta
-    public Optional<ConsultaDTO> findById(Long id) {
-        return consultaRepository.findById(id)
-                .map(ConsultaMapper::convertToDTO);
+
+    public ConsultaDTO findById(Long id) {
+        Consulta consulta = consultaRepository.findById(id)
+                .orElseThrow(() -> new ConsultaNotFoundException("Consulta não encontrada para o id " + id));
+        return ConsultaMapper.convertToDTO(consulta);
     }
 
     @Autowired
@@ -37,6 +39,7 @@ public class ConsultaService {
     private ValidaConflitoHorario validaConflitoHorario;
 
     // Salvar a consulta
+
     public ConsultaDTO save(ConsultaDTO consultaDTO) {
         try {
             validaPaciente.setProximo(validaConflitoHorario);
@@ -50,7 +53,11 @@ public class ConsultaService {
     }
 
     // Deletar com base no ID
+
     public void deleteById(Long id) {
+        if (!consultaRepository.existsById(id)) {
+            throw new ConsultaNotFoundException("Consulta não encontrada para o id " + id);
+        }
         consultaRepository.deleteById(id);
     }
 }
